save success of compiling Will draft

+++ contracts/IOracle.sol
@@ -7,14 +7,14 @@ interface IOracle {
 
     function condition() external view returns (bool);
     function name() external view returns (string memory);
-    function conditionCounter() external view returns (uint);
-    function trustees(uint) external view returns (address);
-    function numerator() external view returns (uint);
-    function denominator() external view returns (uint);
+    function conditionCounter() external view returns (uint256);
+    function trustees(uint256) external view returns (address);
+    function numerator() external view returns (uint256);
+    function denominator() external view returns (uint256);
 
-    function getTrusteeIds(address _trustee) external view returns (uint[] memory);
-    function trusteeOpinion(uint trusteeId) external view returns (bool);
+    function getTrusteeIds(address _trustee) external view returns (uint256[] memory);
+    function trusteeOpinion(uint256 trusteeId) external view returns (bool);
 
-    function setTrustees(address[] memory _trustees, uint _numerator) external;
-    function judge(bool TF, uint trusteeId) external;   
+    function setTrustees(address[] memory _trustees, uint256 _numerator) external;
+    function judge(bool TF, uint256 trusteeId) external;   
 }
